+++ form/src/test/java/com/ritense/form/BaseTest.java
@@ -16,19 +16,37 @@
 
 package com.ritense.form;
 
+import com.ritense.document.config.SpringContextHelper;
 import com.ritense.document.domain.impl.JsonSchema;
 import com.ritense.document.domain.impl.JsonSchemaDocumentDefinition;
 import com.ritense.document.domain.impl.JsonSchemaDocumentDefinitionId;
 import com.ritense.form.domain.FormIoFormDefinition;
+import com.ritense.valtimo.contract.form.FormFieldDataResolver;
 import org.apache.commons.io.IOUtils;
+import org.springframework.context.ApplicationContext;
+
 import java.io.IOException;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
+import java.util.Map;
 import java.util.UUID;
+
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public abstract class BaseTest {
 
     protected static final String DEFAULT_FORM_DEFINITION_NAME = "Form1";
+
+    protected static void mockSpringContextHelper() {
+        var applicationContext = mock(ApplicationContext.class);
+        var formFieldDataResolver = mock(FormFieldDataResolver.class);
+        when(formFieldDataResolver.supports(eq("oz"))).thenReturn(true);
+        when(applicationContext.getBeansOfType(FormFieldDataResolver.class)).thenReturn(Map.of("Test", formFieldDataResolver));
+        var springContextHelper = new SpringContextHelper();
+        springContextHelper.setApplicationContext(applicationContext);
+    }
 
     protected FormIoFormDefinition formDefinition() {
         return new FormIoFormDefinition(UUID.randomUUID(), DEFAULT_FORM_DEFINITION_NAME, "{}", false);
@@ -36,13 +54,6 @@ public abstract class BaseTest {
 
     protected FormIoFormDefinition formDefinition(UUID id, String formName) {
         return new FormIoFormDefinition(id, formName, "{}", false);
-    }
-
-    protected Throwable getRootCause(Throwable throwable) {
-        while (throwable.getCause() != null) {
-            throwable = throwable.getCause();
-        }
-        return throwable;
     }
 
     protected FormIoFormDefinition formDefinitionOf(String formDefinitionId) throws IOException {
